third_party/blink/web_tests/external/wpt/cookies/domain/domain-attribute-idn-host.sub.https-expected.txt

This is a testharness.js-based test.
[FAIL] UTF8-encoded IDN in domain attribute
  assert_equals: expected "no cookies" but got "utf8=set"
[FAIL] UTF8-encoded IDN with non-ASCII dot in domain attribute
  assert_equals: expected "no cookies" but got "utf8-dot=set"
[PASS] wrong UTF8-encoded IDN in domain attribute
[PASS] punycode IDN in domain attribute
[PASS] wrong punycode IDN in domain attribute
[PASS] IDN with invalid UTF-8 bytes in domain attribute
Harness: the test ran to completion.
